Update mxml-get.c
My feeling is that if you were directly accessing the mxml structure for 'type' lets say. It would be a programmer error to dereference a null structure. Now that the accessors are functions I don't think that should change. If the acessors were c++ methods they would not check for a null 'this', and again it would be a programmer error if the method was called on a null object. Ultimately if the function does not crash when null is passed then the user of the library will think that null is a valid argument to a 'get' function.